pyproject.toml
[project]
name = "rover-card"
version = "0.1.0"
description = "Rover Card"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "pillow>=11.3.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
]
